+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Fixed [README.md]
+- Fixed [Get-Ticket] synopsis
+- Added [Get-Ticket] `-ForceXonCreationDateBeforeToday` option
+- Added [Get-Ticket] `-GetSeries` option
 
 ## [6.6.0] - 2023-12-03